requirements.txt
streamlit>=1.30.0
deepface>=0.0.89
chromadb>=0.4.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
plotly>=5.18.0
pandas>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
tf-keras>=2.15.0
